+++ app/assets/javascripts/pwb/shared/vue-main.js
@@ -1,0 +1,6 @@
+window.onload = function() {
+  var pwbSS = Vue.component('social-sharing', SocialSharing);
+  var pwbVue = new Vue({
+    el: '#main-vue'
+  });
+}